# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GeoStrom AI",
    "description": "Retrospective tropical cyclone research prototype. Serves precomputed Phase 2 baseline model predictions. Not an operational forecasting system.",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/v1/cyclones": {
      "get": {
        "tags": [
          "cyclones"
        ],
        "summary": "List Cyclones",
        "operationId": "list_cyclones_api_v1_cyclones_get",
        "parameters": [
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Season"
            }
          },
          {
            "name": "basin",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Basin"
            }
          },
          {
            "name": "split",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "train | val | test",
              "title": "Split"
            },
            "description": "train | val | test"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "substring match on storm id",
              "title": "Q"
            },
            "description": "substring match on storm id"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_StormSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cyclones/{sid}": {
      "get": {
        "tags": [
          "cyclones"
        ],
        "summary": "Get Cyclone",
        "operationId": "get_cyclone_api_v1_cyclones__sid__get",
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StormDetail"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cyclones/{sid}/observations": {
      "get": {
        "tags": [
          "cyclones"
        ],
        "summary": "Get Cyclone Observations",
        "operationId": "get_cyclone_observations_api_v1_cyclones__sid__observations_get",
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "synoptic_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Synoptic Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ObservationOut"
                  },
                  "title": "Response Get Cyclone Observations Api V1 Cyclones  Sid  Observations Get"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tracks/{sid}": {
      "get": {
        "tags": [
          "tracks"
        ],
        "summary": "Get Track",
        "operationId": "get_track_api_v1_tracks__sid__get",
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sid"
            }
          },
          {
            "name": "t",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Forecast origin timestamp (ISO 8601). Defaults to the storm's latest available origin.",
              "title": "T"
            },
            "description": "Forecast origin timestamp (ISO 8601). Defaults to the storm's latest available origin."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureCollection"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prediction/{sid}": {
      "get": {
        "tags": [
          "prediction"
        ],
        "summary": "Get Prediction",
        "operationId": "get_prediction_api_v1_prediction__sid__get",
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sid"
            }
          },
          {
            "name": "t",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Forecast origin timestamp. Defaults to the latest available.",
              "title": "T"
            },
            "description": "Forecast origin timestamp. Defaults to the latest available."
          },
          {
            "name": "task",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "'track' or 'intensity'",
              "title": "Task"
            },
            "description": "'track' or 'intensity'"
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Model name, e.g. 'track_cliper'",
              "title": "Model"
            },
            "description": "Model name, e.g. 'track_cliper'"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PredictionOut"
                  },
                  "title": "Response Get Prediction Api V1 Prediction  Sid  Get"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prediction/{sid}/series": {
      "get": {
        "tags": [
          "prediction"
        ],
        "summary": "Get Prediction Series",
        "description": "Every forecast issued across the storm's life -- powers the\nerror-growth chart (docs/API_ARCHITECTURE.md \u00a73.5).",
        "operationId": "get_prediction_series_api_v1_prediction__sid__series_get",
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sid"
            }
          },
          {
            "name": "lead_hours",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lead Hours"
            }
          },
          {
            "name": "task",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task"
            }
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PredictionOut"
                  },
                  "title": "Response Get Prediction Series Api V1 Prediction  Sid  Series Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prediction/models/list": {
      "get": {
        "tags": [
          "prediction"
        ],
        "summary": "List Models",
        "operationId": "list_models_api_v1_prediction_models_list_get",
        "parameters": [
          {
            "name": "task",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModelVersionOut"
                  },
                  "title": "Response List Models Api V1 Prediction Models List Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Feature": {
        "properties": {
          "type": {
            "type": "string",
            "const": "Feature",
            "title": "Type",
            "default": "Feature"
          },
          "geometry": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PointGeometry"
              },
              {
                "$ref": "#/components/schemas/LineStringGeometry"
              }
            ],
            "title": "Geometry"
          },
          "properties": {
            "additionalProperties": true,
            "type": "object",
            "title": "Properties"
          }
        },
        "type": "object",
        "required": [
          "geometry"
        ],
        "title": "Feature"
      },
      "FeatureCollection": {
        "properties": {
          "type": {
            "type": "string",
            "const": "FeatureCollection",
            "title": "Type",
            "default": "FeatureCollection"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/Feature"
            },
            "type": "array",
            "title": "Features"
          }
        },
        "type": "object",
        "required": [
          "features"
        ],
        "title": "FeatureCollection"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelVersionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "task": {
            "type": "string",
            "title": "Task",
            "description": "'track' or 'intensity'"
          },
          "dataset_build": {
            "type": "string",
            "title": "Dataset Build"
          },
          "metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metrics",
            "description": "Verbatim Phase 2 benchmark metrics, keyed by horizon"
          },
          "error_radii_km": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Radii Km",
            "description": "Track models only: mean historical error (km) per horizon, for rendering an uncertainty envelope"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "version",
          "task",
          "dataset_build",
          "metrics",
          "is_active"
        ],
        "title": "ModelVersionOut"
      },
      "ObservationOut": {
        "properties": {
          "ts": {
            "type": "string",
            "format": "date-time",
            "title": "Ts"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "title": "Lon"
          },
          "wind_kt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Kt"
          },
          "pressure_hpa": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pressure Hpa"
          },
          "category": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "nature": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nature"
          },
          "storm_speed_kt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storm Speed Kt"
          },
          "storm_dir_deg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storm Dir Deg"
          },
          "dist2land_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dist2Land Km"
          },
          "is_synoptic": {
            "type": "boolean",
            "title": "Is Synoptic"
          },
          "is_observed": {
            "type": "boolean",
            "title": "Is Observed"
          },
          "data_kind": {
            "type": "string",
            "title": "Data Kind",
            "default": "observed"
          }
        },
        "type": "object",
        "required": [
          "ts",
          "lat",
          "lon",
          "wind_kt",
          "pressure_hpa",
          "category",
          "nature",
          "storm_speed_kt",
          "storm_dir_deg",
          "dist2land_km",
          "is_synoptic",
          "is_observed"
        ],
        "title": "ObservationOut"
      },
      "Page_StormSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StormSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[StormSummary]"
      },
      "PointGeometry": {
        "properties": {
          "type": {
            "type": "string",
            "const": "Point",
            "title": "Type",
            "default": "Point"
          },
          "coordinates": {
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Coordinates",
            "description": "[lon, lat]"
          }
        },
        "type": "object",
        "required": [
          "coordinates"
        ],
        "title": "PointGeometry"
      },
...
      "ProblemDetail": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "about:blank"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          }
        },
        "type": "object",
        "required": [
          "title",
          "status"
        ],
        "title": "ProblemDetail",
        "description": "RFC 7807 problem+json, per docs/API_ARCHITECTURE.md \u00a72."
      },
      "StormDetail": {
        "properties": {
          "sid": {
            "type": "string",
            "title": "Sid"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "season": {
            "type": "integer",
            "title": "Season"
          },
          "basin": {
            "type": "string",
            "title": "Basin"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "title": "End Time"
          },
          "n_observations": {
            "type": "integer",
            "title": "N Observations"
          },
          "max_wind_kt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Wind Kt",
            "description": "Lifetime max wind (kt). DISPLAY ONLY."
          },
          "min_pressure_hpa": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Pressure Hpa"
          },
          "max_category": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Category"
          },
          "made_landfall": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Made Landfall"
          },
          "split": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Split",
            "description": "train | val | test (Phase 2 frozen split)"
          },
          "has_predictions": {
            "type": "boolean",
            "title": "Has Predictions"
          },
          "bbox": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bbox",
            "description": "[min_lon, min_lat, max_lon, max_lat]"
          }
        },
        "type": "object",
        "required": [
          "sid",
          "name",
          "season",
          "basin",
          "start_time",
          "end_time",
          "n_observations",
          "min_pressure_hpa",
          "max_category",
          "made_landfall",
          "has_predictions"
        ],
        "title": "StormDetail"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}